requirements.django.txt
Django==1.11
django-payments==0.12.3
jsonfield
mock==2.0.0
pyOpenSSL==18.0.0
